% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adestr_package.R
\docType{package}
\name{adestr}
\alias{adestr}
\alias{adestr-package}
\title{adestr}
\description{
Point estimates, confidence intervals, and p-values for optimal adaptive two-stage designs
}
\details{
This package implements methods to evaluate the performance characteristics of
various point and interval estimators for optimal adaptive two-stage designs.
Specifically, this package is written to interface with trial designs created by the \code{\link{adoptr}} package
\insertCite{kunzmann2021adoptr,pilz2021optimal}{adestr}.
Apart from the a priori evaluation of performance characteristics, this package also allows for the
evaluation of the implemented estimators on real datasets, and it implements methods
to calculate p-values.
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jan-imbi.github.io/adestr/}
}

}
\author{
\strong{Maintainer}: Jan Meis \email{meis@imbi.uni-heidelberg.de} (\href{https://orcid.org/0000-0001-5407-7220}{ORCID})

}
