% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_estimator.R
\name{c,EstimatorScoreResult-method}
\alias{c,EstimatorScoreResult-method}
\title{Combine EstimatoreScoreResult objects into a list}
\usage{
\S4method{c}{EstimatorScoreResult}(x, ...)
}
\arguments{
\item{x}{an object of class EstimatorScoreResult.}

\item{...}{additional arguments passed along to the \code{\link{list}} function}
}
\value{
an object of class EstimatoreScoreResultList.
}
\description{
Creates an object of class EstimatoreScoreResultList,
which is a basically list with the respective
EstimatoreScoreResult objects.
}
