% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2c2_helpers.R
\name{n2_extrapol}
\alias{n2_extrapol}
\title{Calculate the second-stage sample size for a design with cached spline parameters}
\usage{
n2_extrapol(design, x1)
}
\arguments{
\item{design}{an object of class \code{\link{TwoStageDesignWithCache}}.}

\item{x1}{first-stage test statistic}
}
\description{
Also extrapolates results for values outside of [c1f, c1e].
}
