% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\docType{class}
\name{PointEstimator-class}
\alias{PointEstimator-class}
\alias{PointEstimator}
\alias{VirtualPointEstimator}
\alias{SampleMean}
\alias{FirstStageSampleMean}
\alias{WeightedSampleMean}
\alias{AdaptivelyWeightedSampleMean}
\alias{MinimizePeakVariance}
\alias{BiasReduced}
\alias{RaoBlackwell}
\alias{PseudoRaoBlackwell}
\alias{MidpointStagewiseCombinationFunctionOrderingCI}
\alias{MidpointMLEOrderingCI}
\alias{MidpointLikelihoodRatioOrderingCI}
\alias{MidpointScoreTestOrderingCI}
\alias{MidpointNeymanPearsonOrderingCI}
\alias{MedianUnbiasedStagewiseCombinationFunctionOrdering}
\alias{MedianUnbiasedMLEOrdering}
\alias{MedianUnbiasedLikelihoodRatioOrdering}
\alias{MedianUnbiasedScoreTestOrdering}
\alias{MedianUnbiasedNeymanPearsonOrdering}
\title{Point estimators}
\usage{
PointEstimator(g1, g2, label)

VirtualPointEstimator()

SampleMean()

FirstStageSampleMean()

WeightedSampleMean(w1 = 0.5)

AdaptivelyWeightedSampleMean(w1 = 1/sqrt(2))

MinimizePeakVariance()

BiasReduced(iterations = 1L)

RaoBlackwell()

PseudoRaoBlackwell()

MidpointStagewiseCombinationFunctionOrderingCI()

MidpointMLEOrderingCI()

MidpointLikelihoodRatioOrderingCI()

MidpointScoreTestOrderingCI()

MidpointNeymanPearsonOrderingCI()

MedianUnbiasedStagewiseCombinationFunctionOrdering()

MedianUnbiasedMLEOrdering()

MedianUnbiasedLikelihoodRatioOrdering()

MedianUnbiasedScoreTestOrdering()

MedianUnbiasedNeymanPearsonOrdering(mu0 = 0, mu1 = 0.4)
}
\arguments{
\item{g1}{functional representation of the estimator in the early futility and efficacy regions.}

\item{g2}{functional representation of the estimator in the continuation region.}

\item{label}{name of the estimator. Used in printing methods.}

\item{w1}{weight of the first-stage sample mean.}

\item{iterations}{number of bias reduction iterations. Defaults to 1.}

\item{mu0}{expected value of the normal distribution under the null hypothesis.}

\item{mu1}{expected value of the normal distribution under the null hypothesis.}
}
\value{
An object of class \code{PointEstimator}.
}
\description{
Point estimators
}
\examples{
PointEstimator(g1 = \(smean1, ...) smean1,g2 = \(smean2, ...) smean2, label="My custom estimator")
}
