% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostagedesign_with_cache.R
\docType{class}
\name{TwoStageDesign-class}
\alias{TwoStageDesign-class}
\title{Re-export of two-stage design class}
\description{
This is a re-export of the \code{TwoStageDesign} class from the
\code{adoptr} \insertCite{kunzmann2021adoptr}{adestr} package.
}
\details{
This function is currently re-exported here to resolve CRAN conflicts.
For details, please refer to the paper \insertCite{kunzmann2021adoptr}{adestr}
and the original github repository \url{https://github.com/kkmann/adoptr}.
}
\section{Slots}{

\describe{
\item{\code{n1}}{(numeric) first-stage sample size.}

\item{\code{c1f}}{(numeric) first-stage futility boundary.}

\item{\code{c1e}}{(numeric) first-stage early efficacy boundary.}

\item{\code{n2_pivots}}{(numeric) vector containing the values of the n2 spline function.}

\item{\code{c2_pivots}}{(numeric) vector containing the values of the second-stage
rejection boundary spline c2}

\item{\code{x1_norm_pivots}}{(numeric) vector containing the x-axis (z-sclae) points
for the n2 and c2 splines}

\item{\code{weights}}{(numeric) vector containing integration weights}

\item{\code{tunable}}{(logical) vector determining whether desing paramters are to be optimized}
}}

\seealso{
The original implementation of the adoptr package by Kevin Kunzmann and
Maximilian Pilz is available at \url{https://github.com/kkmann/adoptr}.
}
