% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostagedesign_with_cache.R
\name{TwoStageDesignWithCache}
\alias{TwoStageDesignWithCache}
\title{TwoStageDesignWithCache constructor function}
\usage{
TwoStageDesignWithCache(design)
}
\arguments{
\item{design}{an object of class TwoStageDesign}
}
\description{
Creates an object of class \code{TwoStageDesignWithCache}.
This object stores the precalculated spline paramters of the \code{n2}
and \code{c2} functions, which allows for quicker evaluation.
}
