% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{UniformPrior}
\alias{UniformPrior}
\title{Uniform prior distribution for the parameter mu}
\usage{
UniformPrior(min = -1, max = 1)
}
\arguments{
\item{min}{minimum of support interval.}

\item{max}{maximum of support interval.}
}
\value{
an object of class \code{UniformPrior}. This object can be supplied
as the argument \code{mu} of the \code{\link{evaluate_estimator}} function
to calculate performance scores weighted by a prior.
}
\description{
Uniform prior distribution for the parameter mu
}
\examples{
UniformPrior(min = -1, max = 1)
}
