% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_estimator.R
\docType{class}
\name{EstimatorScore-class}
\alias{EstimatorScore-class}
\alias{EstimatorScore}
\alias{Expectation}
\alias{Bias}
\alias{Variance}
\alias{MSE}
\alias{OverestimationProbability}
\alias{Coverage}
\alias{SoftCoverage}
\alias{Width}
\alias{TestAgreement}
\title{Performance scores for point and interval estimators}
\usage{
Expectation()

Bias()

Variance()

MSE()

OverestimationProbability()

Coverage()

SoftCoverage(shrinkage = 1)

Width()

TestAgreement()
}
\arguments{
\item{shrinkage}{shrinkage factor for bump function.}
}
\value{
an \code{EstimatorScore} object.
}
\description{
These classes encode various metrics which can be used to evaluate
the performance characteristics of point and interval estimators.
}
\section{Slots}{

\describe{
\item{\code{label}}{name of the performance score. Used in printing methods.}
}}

\examples{
evaluate_estimator(
  score = MSE(),
  estimator = SampleMean(),
  data_distribution = Normal(FALSE),
  design = get_example_design(),
  mu = c(0.3),
  sigma = 1,
  exact = FALSE
)
}
\seealso{
\code{\link{evaluate_estimator}}
}
