% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_estimator.R
\docType{class}
\name{EstimatorScore-class}
\alias{EstimatorScore-class}
\alias{EstimatorScore}
\alias{Expectation}
\alias{Bias}
\alias{Variance}
\alias{MSE}
\alias{OverestimationProbability}
\alias{Coverage}
\alias{SoftCoverage}
\alias{Width}
\alias{TestAgreement}
\alias{Centrality}
\title{Performance scores for point and interval estimators}
\usage{
Expectation()

Bias()

Variance()

MSE()

OverestimationProbability()

Coverage()

SoftCoverage(shrinkage = 1)

Width()

TestAgreement()

Centrality(interval = NULL)
}
\arguments{
\item{shrinkage}{shrinkage factor for bump function.}
}
\value{
an \code{EstimatorScore} object.
}
\description{
These classes encode various metrics which can be used to evaluate
the performance characteristics of point and interval estimators.
}
\section{Slots}{

\describe{
\item{\code{label}}{name of the performance score. Used in printing methods.}
}}

\section{Details on the implemented estimators}{
In the following, precise definitions of the performance scores implemented
in \code{\link{adestr}}
are given. To this end,
let \eqn{\hat{\mu}} denote a point estimator, (\eqn{\hat{l}}, \eqn{\hat{u}})
an interval estimator, denote the expected value of a random variable
by \eqn{\mathbb{E}}, and let \eqn{\mu} be the real value of the underlying
parameter to be estimated.
\subsection{Scores for point estimators (\code{PointEstimatorScore}):}{
\itemize{
\item \code{Expectation()}: \eqn{\mathbb{E}[\hat{\mu}]}
\item \code{Bias()}: \eqn{\mathbb{E}[\hat{\mu} - \mu]}
\item \code{Variance()}: \eqn{\mathbb{E}[(\hat{\mu} - \mathbb{E}[\hat{\mu}])^2]}
\item \code{MSE()}: \eqn{\mathbb{E}[(\hat{\mu} - mu)^2]}
}
}
}

\examples{
evaluate_estimator(
  score = MSE(),
  estimator = SampleMean(),
  data_distribution = Normal(FALSE),
  design = get_example_design(),
  mu = c(0, 0.3, 0.6),
  sigma = 1,
  exact = FALSE
)
}
\seealso{
\code{\link{evaluate_estimator}}
}
