% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_estimator.R
\name{evaluate_estimator}
\alias{evaluate_estimator}
\title{Evaluate performance characteristics of an estimator}
\usage{
evaluate_estimator(
  score,
  estimator,
  data_distribution,
  use_full_twoarm_sampling_distribution = FALSE,
  design,
  true_parameter = mu,
  mu,
  sigma,
  tol = getOption("adestr_tol_outer", default = .adestr_options[["adestr_tol_outer"]]),
  maxEval = getOption("adestr_maxEval_outer", default =
    .adestr_options[["adestr_maxEval_outer"]]),
  absError = getOption("adestr_absError_outer", default =
    .adestr_options[["adestr_absError_outer"]]),
  exact = FALSE,
  early_futility_part = TRUE,
  continuation_part = TRUE,
  early_efficacy_part = TRUE,
  conditional_integral = FALSE
)
}
\arguments{
\item{score}{performance measure to evaluate.}

\item{estimator}{object of class \code{PointEstimator}, \code{IntervalEstimator} or \code{PValue}.}

\item{data_distribution}{object of class \code{Normal} or \code{Student}.}

\item{use_full_twoarm_sampling_distribution}{logical indicating whether this estimator is intended to be used
with the full sampling distribution in a two-armed trial.}

\item{design}{object of class \code{TwoStageDesign}.}

\item{true_parameter}{true value of the parameter (used e.g. when evaluating bias).}

\item{mu}{expected value of the underlying normal distribution.}

\item{sigma}{assumed standard deviation.}

\item{tol}{relative tolerance.}

\item{maxEval}{maximum number of iterations.}

\item{absError}{absolute tolerance.}

\item{exact}{logical indicating usage of exact n2 function.}

\item{early_futility_part}{include early futility part of integral.}

\item{continuation_part}{include continuation part of integral.}

\item{early_efficacy_part}{include early efficacy part of integral.}

\item{conditional_integral}{treat integral as a conditional integral.}
}
\value{
\code{EstimatorScoreResult} object containing the performance characteristics of the estimator.
}
\description{
Evaluate performance characteristics of an estimator
}
\examples{
evaluate_estimator(
  score = MSE(),
  estimator = SampleMean(),
  data_distribution = Normal(FALSE),
  design = get_example_design(),
  mu = c(0, 0.3, 0.6),
  sigma = 1,
  exact = FALSE
)
}
