% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_example_design}
\alias{get_example_design}
\title{Generate an exemplary adaptive design}
\usage{
get_example_design(label = NULL)
}
\value{
an exmplary design of class \code{TwoStageDesign}.
}
\description{
The design was optimized to minimize the expected sample size
under the alternative hypothesis for a one-armed trial.
The boundaries are chosen to control the type I error at 0.025
for a normally distributed test statistic (i.e. known variance).
For an alternative hypothesis of mu=0.4, the overall power is 80\%.
}
\examples{
design <- get_example_design()
# Type I error
evaluate(Power(Normal(FALSE), PointMassPrior(0, 1)), design)
# Power
evaluate(Power(Normal(FALSE), PointMassPrior(.4, 1)), design)
# Expected sample size under the alternative
evaluate(ExpectedSampleSize(Normal(FALSE), PointMassPrior(.4, 1)), design)
# Expected sample size under the null
evaluate(ExpectedSampleSize(Normal(FALSE), PointMassPrior(0, 1)), design)
}
