% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_statistics_from_paper}
\alias{get_statistics_from_paper}
\title{Generate a list of estimators and p-values to use in examples}
\usage{
get_statistics_from_paper(
  point_estimators = TRUE,
  interval_estimators = TRUE,
  p_values = TRUE
)
}
\arguments{
\item{point_estimators}{logical indicating whether point estimators should be included in output list}

\item{interval_estimators}{logical indicating whether interval estimators should be included in output list}

\item{p_values}{logical indicating whether p-values should be included in output list}
}
\value{
a list of estimators and pvalues.
}
\description{
Generate a list of estimators and p-values to use in examples
}
\examples{
set.seed(321)
dat <- data.frame(
  endpoint = c(rnorm(28, .2, 1), rnorm(28, 0, 1),
               rnorm(23, .2, 1), rnorm(23, 0, 1)),
  group = factor(rep(c("ctl", "trt", "ctl", "trt"),
                     c(28,28,23,23))),
  stage = rep(c(1L, 2L), c(56, 46))
)
analyze(
  data = dat,
  statistics = c(get_example_statistics()),
  data_distribution = Normal(),
  sigma = 1,
  design = get_example_design()
)
}
