% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostagedesign_with_cache.R
\name{Normal}
\alias{Normal}
\title{Normally distributed data with known variance}
\usage{
Normal(two_armed = TRUE)
}
\arguments{
\item{two_armed}{(logical) determines whether one or two-armed trials are assumed.}
}
\description{
This function creates an object representing the distributional assumptions
of the data: normally distributed outcomes sample from a trial with
one or two arms (depending on the value of the parameter \code{two_armed}),
under the assumption of known variance.
}
