% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{NormalPrior}
\alias{NormalPrior}
\title{Normal prior distribution for the parameter mu}
\usage{
NormalPrior(mu = 0, sigma = 1)
}
\arguments{
\item{mu}{mean of prior distribution.}

\item{sigma}{standard deviation of the prior distribution.}
}
\value{
An object of class \code{NormalPrior}.
}
\description{
Normal prior distribution for the parameter mu
}
\examples{
NormalPrior(mu = 0, sigma = 1)
}
