% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R, R/neyman_pearson.R
\docType{class}
\name{PValue-class}
\alias{PValue-class}
\alias{PValue}
\alias{VirtualPValue}
\alias{LinearShiftRepeatedPValue}
\alias{MLEOrderingPValue}
\alias{LikelihoodRatioOrderingPValue}
\alias{ScoreTestOrderingPValue}
\alias{StagewiseCombinationFunctionOrderingPValue}
\alias{NeymanPearsonOrderingPValue}
\title{P-values}
\usage{
PValue(g1, g2, label)

VirtualPValue()

LinearShiftRepeatedPValue(wc1f = 0, wc1e = 1/2, wc2 = 1/2)

MLEOrderingPValue()

LikelihoodRatioOrderingPValue()

ScoreTestOrderingPValue()

StagewiseCombinationFunctionOrderingPValue()

NeymanPearsonOrderingPValue(mu0 = 0, mu1 = 0.4)
}
\arguments{
\item{g1}{functional representation of the p-value in the early futility and efficacy regions.}

\item{g2}{functional representation of the p-value in the continuation region.}

\item{label}{name of the p-value. Used in printing methods.}

\item{wc1f}{slope of futility boundary change.}

\item{wc1e}{slope of efficacy boundary change.}

\item{wc2}{slope of c2 boundary change.}

\item{mu0}{expected value of the normal distribution under the null hypothesis.}

\item{mu1}{expected value of the normal distribution under the null hypothesis.}
}
\value{
An object of class \code{PValue}.
}
\description{
P-values
}
\examples{
PValue(
  g1 = \(smean1, ...) runif(length(smean1)),
  g2 = \(smean2, ...) runif(length(smean2)),
  label="My custom p-value")
}
